backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.2.0
Django==4.1.4
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
finviz @ git+https://github.com/mariostoev/finviz@83de213e996a6672a0b395a6cf55e18d3728a675
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
multidict==6.0.3
multitasking==0.0.11
numpy==1.23.5
pandas==1.5.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==7.0.0
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
user-agent==0.1.10
vine==5.0.0
wcwidth==0.2.5
yarl==1.8.2
yfinance==0.1.90
